# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deve Sub API",
    "description": "Self-hosted proxy subscription infrastructure manager",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/nodes/batch-enabled": {
      "post": {
        "summary": "`POST /api/v1/nodes/batch-enabled` — batch set enabled flag (NODE-004).",
        "operationId": "batch_set_enabled",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid node ids",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Node or tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/nodes/batch-tags": {
      "post": {
        "summary": "`POST /api/v1/nodes/batch-tags` — batch replace tags (NODE-005).",
        "operationId": "batch_set_tags",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTagsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Batch tags applied"
          },
          "400": {
            "description": "Invalid ids",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Node or tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/v1/nodes/{id}/chain": {
      "put": {
        "summary": "`PUT /api/v1/nodes/{id}/chain` — set or clear a node's proxy chain\n(NODE-017 / NODE-018, admin only).",
        "operationId": "set_node_chain",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node ULID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetNodeChainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chain updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeChainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid node id or chain structure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Node not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Chain would create a cycle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
    "/api/v1/nodes/{id}/region": {
      "patch": {
        "summary": "`PATCH /api/v1/nodes/{id}/region` — set or clear a manual region (NODE-006).",
        "operationId": "set_region",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node ULID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRegionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Region updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid node id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Node not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/nodes/{id}/tags": {
      "put": {
        "summary": "`PUT /api/v1/nodes/{id}/tags` — replace tags for a single node (NODE-005).",
        "operationId": "set_node_tags",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node ULID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetNodeTagsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Tags updated"
          },
          "400": {
            "description": "Invalid ids",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Node or tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchEnabledRequest": {
        "type": "object",
        "description": "Request body for `POST /api/v1/nodes/batch-enabled` (NODE-004).",
        "required": [
          "node_ids",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "`true` to enable, `false` to disable."
          },
          "node_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Node ULIDs to update."
          }
        },
        "additionalProperties": false
      },
      "BatchResultDto": {
        "type": "object",
        "description": "Response body for batch operations.",
        "required": [
          "updated"
        ],
        "properties": {
          "updated": {
            "type": "integer",
            "format": "int64",
            "description": "Number of rows affected.",
            "minimum": 0
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response for auth endpoints.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `\"unauthorized\"`, `\"forbidden\"`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "NodeChainResponse": {
        "type": "object",
        "description": "Response body for chain endpoints.",
        "required": [
          "nodes"
        ],
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The node IDs forming the chain. Empty means direct connection."
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "deve_sub_session"
      }
    }
  }
}